+++ examples/spot/pay/payHistory.js
@@ -7,7 +7,7 @@ const client = new Spot(apiKey, apiSecret)
 
 client.payHistory(
   {
-    startTimestamp: 1637570276000
+    startTime: 1637570276000
   }
 ).then(response => client.logger.log(response.data))
   .catch(error => client.logger.error(error))